Move get trainer to all role in system Add remove notyet trainerslot for unhappy case check out birdtrainingcourse Add remove all trainerslot for unhappy case cancel birdtrainingcourse Add br condition training slot must be done in right date
